requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
Django==2.2.6
django-rest-knox==4.1.0
djangorestframework==3.10.3
fake-headers==1.0.2
html5lib==1.0.1
idna==2.8
Pillow==6.2.0
psycopg2-binary==2.8.3
pycparser==2.19
pytz==2019.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
urllib3==1.25.6
webencodings==0.5.1
